feat(og): support forwarding headers
Previously, the `Host` header would not be forwarded which could cause
issues with the application that a request was being sent to. This fixes
that.

Also included is `X-Forwarded-Proto` for the case of being behind a
reverse proxy, such as `Cloudflare`.

This is not a perfect or elegant solution and is meant to solve a few
specific use cases. We should consider something better, such as always
forwarding headers in the future.

Note: I re-ran `npm install` locally which updated `package-lock.json`