Fail immediately if pass-challenge is sent without cookie
This is a straightforward way to test for cookie support in the user
agent. It doesn't change the behavior for any other user agents, but
browsers that have cookies disabled will show an error rather than
refreshing endlessly.

This fixes a common user complaint - #428 and also discussed on HN:
https://news.ycombinator.com/item?id=43865151

Most users don't block cookies entirely in their browser, but this
would help the few that do, and for users selectively enabling cookies
for certain sites, this will make allow-listing Anubis fronted sites
easier because the page will not be redirecting while they're trying
to allow the cookie to be set.